refactor: 이벤트 API refactor
